chore: renames summarize to metrics dashboard
